requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
discord==1.7.3
discord.py==1.7.3
idna==3.2
multidict==5.1.0
Pillow==8.3.1
requests==2.26.0
selenium==3.141.0
typing-extensions==3.10.0.0
urllib3==1.26.6
webdriver-manager==3.4.2
yarl==1.6.3
